django_project/requirements.txt
django==4.2.2
gunicorn==20.1.0
psycopg==3.1.9
django-simple-history==3.3.0
drf-yasg==1.21.6
django-filter==23.2
djangorestframework==3.14.0
python-dotenv==1.0.0
coreapi==2.3.3